+++ tests/testthat/test-mtcars-ranger.R
@@ -108,7 +108,7 @@ test_that("ranger: constrained decomposition invalidates shap", {
     "efficiency property"
   )
   expect_identical(gl_mi$constrained, "max_interaction")
-  expect_true(all(is.na(gl_mi$shap)))
+  expect_identical(gl_mi$shap, NA)
   expect_false(anyNA(gl_mi$m))
 
   expect_warning(
@@ -116,7 +116,7 @@ test_that("ranger: constrained decomposition invalidates shap", {
     "efficiency property"
   )
   expect_identical(gl_ft$constrained, "features")
-  expect_true(all(is.na(gl_ft$shap)))
+  expect_identical(gl_ft$shap, NA)
   expect_false(anyNA(gl_ft$m))
 })